Refactor RenaperRoutes MM-310
