+++ controllers/renaper/index.js
@@ -1,0 +1,7 @@
+const { validateDni } = require('./validateDni');
+const { validateDniState } = require('./validateDniState');
+
+module.exports = {
+  validateDni,
+  validateDniState,
+}
